+++ Mile.Helpers/Mile.Helpers.cpp
@@ -288,3 +288,9 @@ EXTERN_C BOOL WINAPI MileAllowDarkModeForApp(
 
     return false;
 }
+
+EXTERN_C COLORREF WINAPI MileGetDefaultBackgroundColorValue(
+    _In_ BOOL UseImmersiveDarkMode)
+{
+    return (UseImmersiveDarkMode ? RGB(32, 32, 32) : RGB(243, 243, 243));
+}