+++ ReleaseNotes.md
@@ -14,3 +14,4 @@
 - Add MILE_PREFERRED_APP_MODE enum.
 - Add MileSetPreferredAppMode function.
 - Add MileAllowDarkModeForApp function.
+- Add MileGetDefaultBackgroundColorValue function.